Added a lot more code, to calculate song performance intensity and prepare for possible choice modelling.
Song performance intensity = number of times a song was played / number of times available as an option.  

A song is considered available in the repertoire from the first time it is performed.  A song is considered available during a gig if it is available in the repertoire and if it has not already been played during that gig.